Replace drawings in TangleStore with wrapped vivagraphLib
